+++ weblate/lang/models.py
@@ -121,8 +121,6 @@ def get_default_lang():
 
 
 class LanguageQuerySet(models.QuerySet):
-    # pylint: disable=no-init, 1000000
-
     def try_get(self, *args, **kwargs):
         """Try to get language by code."""
         result = self.filter(*args, **kwargs)[:2]
@@ -176,7 +174,7 @@ class LanguageQuerySet(models.QuerySet):
         # Strip leading and trailing .
         code = code.strip(".")
 
-        return code  # noqa: RET504
+        return code
 
     def aliases_get(self, code, expanded_code=None):
         code = code.lower()
@@ -902,7 +900,7 @@ class PluralMapper:
         obj = cls.instances.get(key)
         if obj is None:
             obj = cls.instances[key] = super().__new__(cls)
-        return obj  # noqa: RET504
+        return obj
 
     def __init__(self, source_plural, target_plural):
         self.source_plural = source_plural