+++ weblate/metrics/tasks.py
@@ -2,10 +2,9 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from datetime import timedelta
+from datetime import date, timedelta
 
 from celery.schedules import crontab
-from django.utils import timezone
 
 from weblate.auth.models import User
 from weblate.lang.models import Language
@@ -35,33 +34,18 @@ def collect_metrics():
 @app.task(trail=False)
 def cleanup_metrics():
     """Remove stale metrics."""
-    # Remove metrics for deleted objects
-    projects = Project.objects.values_list("pk", flat=True)
-    Metric.objects.filter(scope=Metric.SCOPE_PROJECT).exclude(
-        relation__in=projects
-    ).delete()
-    Metric.objects.filter(scope=Metric.SCOPE_PROJECT_LANGUAGE).exclude(
-        relation__in=projects
-    ).delete()
-    Metric.objects.filter(scope=Metric.SCOPE_COMPONENT).exclude(
-        relation__in=Component.objects.values_list("pk", flat=True)
-    ).delete()
-    Metric.objects.filter(scope=Metric.SCOPE_TRANSLATION).exclude(
-        relation__in=Translation.objects.values_list("pk", flat=True)
-    ).delete()
-    Metric.objects.filter(scope=Metric.SCOPE_USER).exclude(
-        relation__in=User.objects.values_list("pk", flat=True)
-    ).delete()
-    Metric.objects.filter(scope=Metric.SCOPE_COMPONENT_LIST).exclude(
-        relation__in=ComponentList.objects.values_list("pk", flat=True)
-    ).delete()
-    Metric.objects.filter(scope=Metric.SCOPE_LANGUAGE).exclude(
-        relation__in=Language.objects.values_list("pk", flat=True)
-    ).delete()
-
     # Remove past metrics, but we need data for last 24 months
-    cutoff = timezone.now() - timedelta(days=800)
-    Metric.objects.filter(date__lte=cutoff).delete()
+    Metric.objects.filter(date__lte=date.today() - timedelta(days=800)).delete()
+
+    # Remove detailed data for past metrics, we need details only for two months
+    # - avoid filtering on data field as that one is not indexed
+    # - wipe only interval of data with assumption that this task is executed daily
+    Metric.objects.filter(
+        date__range=(
+            date.today() - timedelta(days=75),
+            date.today() - timedelta(days=65),
+        )
+    ).update(data=None)
 
 
 @app.on_after_finalize.connect