+++ weblate/trans/discovery.py
@@ -188,7 +188,7 @@ class ComponentDiscovery:
             result = match[key]
             if len(result) > max_length - extra:
                 result = result[: max_length - extra]
-            return result  # noqa: RET504
+            return result
 
         # Get name and slug
         name = get_val("name") or "Component"
@@ -254,9 +254,8 @@ class ComponentDiscovery:
                 # Valid new base?
                 if os.path.exists(component.get_new_base_filename()):
                     continue
-            else:
-                if component.get_mask_matches():
-                    continue
+            elif component.get_mask_matches():
+                continue
 
             # Delete as needed files seem to be missing
             deleted.append((None, component))