+++ weblate/trans/templatetags/translations.py
@@ -58,8 +58,8 @@ GLOSSARY_TEMPLATE = """<span class="glossary-term" title="{}">"""
 # This should match whitespace_regex in weblate/static/loader-bootstrap.js
 WHITESPACE_REGEX = (
     r"(\t|\u00A0|\u1680|\u2000|\u2001|\u2002|\u2003|"
-    + r"\u2004|\u2005|\u2006|\u2007|\u2008|\u2009|\u200A|"
-    + r"\u202F|\u205F|\u3000)"
+    r"\u2004|\u2005|\u2006|\u2007|\u2008|\u2009|\u200A|"
+    r"\u202F|\u205F|\u3000)"
 )
 WHITESPACE_RE = re.compile(WHITESPACE_REGEX, re.MULTILINE)
 MULTISPACE_RE = re.compile(r"(  +| $|^ )", re.MULTILINE)
@@ -414,9 +414,6 @@ def show_message(tags, message):
 
 def naturaltime_past(value, now):
     """Handling of past dates for naturaltime."""
-    # this function is huge
-    # pylint: disable=too-many-branches,too-many-return-statements
-
     delta = now - value
 
     if delta.days >= 365:
@@ -471,9 +468,6 @@ def naturaltime_past(value, now):
 
 def naturaltime_future(value, now):
     """Handling of future dates for naturaltime."""
-    # this function is huge
-    # pylint: disable=too-many-branches,too-many-return-statements
-
     delta = value - now
 
     if delta.days >= 365: