+++ weblate/trans/views/edit.py
@@ -64,6 +64,8 @@ from weblate.utils.views import (
     show_form_errors,
 )
 
+SESSION_SEARCH_CACHE_TTL = 1800
+
 
 def parse_params(request, project, component, lang):
     """Parses base object and unit set from request."""
@@ -101,19 +103,20 @@ def get_other_units(unit):
     component = translation.component
     propagation = component.allow_translation_propagation
     same = None
+    any_propagated = False
 
     if unit.source and unit.context:
         match = Q(source=unit.source) & Q(context=unit.context)
         if component.has_template():
-            query = Q(source__ilike=unit.source) | Q(context__ilike=unit.context)
+            query = Q(source__iexact=unit.source) | Q(context__iexact=unit.context)
         else:
-            query = Q(source__ilike=unit.source)
+            query = Q(source__iexact=unit.source)
     elif unit.source:
         match = Q(source=unit.source) & Q(context="")
-        query = Q(source__ilike=unit.source)
+        query = Q(source__iexact=unit.source)
     elif unit.context:
         match = Q(context=unit.context)
-        query = Q(context__ilike=unit.context)
+        query = Q(context__iexact=unit.context)
     else:
         return result
 
@@ -171,6 +174,8 @@ def get_other_units(unit):
             and item.source == unit.source
             and item.context == unit.context
         )
+        if item.pk != unit.pk:
+            any_propagated |= item.is_propagated
         untranslated |= not item.translated
         allow_merge |= item.allow_merge
         if item.pk == unit.pk:
@@ -187,7 +192,7 @@ def get_other_units(unit):
 
     # Slightly different logic to allow applying current translation to
     # the propagated strings
-    if same is not None:
+    if same is not None and any_propagated:
         same.allow_merge = (
             (untranslated or allow_merge) and same.translated and propagation
         )
@@ -207,12 +212,7 @@ def cleanup_session(session, delete_all: bool = False):
         if not key.startswith("search_"):
             continue
         value = session[key]
-        if (
-            delete_all
-            or not isinstance(value, dict)
-            or value["ttl"] < now
-            or "items" not in value
-        ):
+        if delete_all or not isinstance(value, dict) or value["ttl"] < now:
             del session[key]
 
 
@@ -220,6 +220,7 @@ def search(
     base, project, unit_set, request, blank: bool = False, use_cache: bool = True
 ):
     """Perform search or returns cached search results."""
+    now = int(time.monotonic())
     # Possible new search
     form = PositionSearchForm(user=request.user, data=request.GET, show_builder=False)
 
@@ -245,13 +246,13 @@ def search(
     }
     session_key = f"search_{base.cache_key}_{search_url}"
 
-    if (
-        use_cache
-        and session_key in request.session
-        and "offset" in request.GET
-        and "items" in request.session[session_key]
-    ):
+    # Remove old search results
+    cleanup_session(request.session)
+
+    session_data = request.session.get(session_key)
+    if use_cache and session_data and "offset" in request.GET:
         search_result.update(request.session[session_key])
+        request.session[session_key]["ttl"] = now + SESSION_SEARCH_CACHE_TTL
         return search_result
 
     allunits = unit_set.search(cleaned_data.get("q", ""), project=project)
@@ -266,9 +267,6 @@ def search(
         messages.warning(request, _("No strings found!"))
         return redirect(base)
 
-    # Remove old search results
-    cleanup_session(request.session)
-
     store_result = {
         "query": search_query,
         "url": search_url,
@@ -276,7 +274,7 @@ def search(
         "key": session_key,
         "name": str(name),
         "ids": unit_ids,
-        "ttl": int(time.monotonic()) + 86400,
+        "ttl": now + SESSION_SEARCH_CACHE_TTL,
     }
     if use_cache:
         request.session[session_key] = store_result
@@ -287,7 +285,7 @@ def search(
 
 def perform_suggestion(unit, form, request):
     """Handle suggesion saving."""
-    if form.cleaned_data["target"][0] == "":
+    if not form.cleaned_data["target"][0]:
         messages.error(request, _("Your suggestion is empty!"))
         # Stay on same entry
         return False
@@ -296,11 +294,12 @@ def perform_suggestion(unit, form, request):
         messages.error(request, _("You don't have privileges to add suggestions!"))
         # Stay on same entry
         return False
-    if not request.user.is_authenticated:
-        # Spam check
-        if is_spam("\n".join(form.cleaned_data["target"]), request):
-            messages.error(request, _("Your suggestion has been identified as spam!"))
-            return False
+    # Spam check for unauthenticated users
+    if not request.user.is_authenticated and is_spam(
+        "\n".join(form.cleaned_data["target"]), request
+    ):
+        messages.error(request, _("Your suggestion has been identified as spam!"))
+        return False
 
     # Create the suggestion
     result = Suggestion.objects.add(