+++ weblate/vcs/git.py
@@ -22,6 +22,7 @@ from django.core.cache import cache
 from django.utils.functional import cached_property
 from django.utils.translation import gettext, gettext_lazy
 from git.config import GitConfigParser
+from requests.exceptions import HTTPError
 
 from weblate.utils.data import data_dir
 from weblate.utils.errors import report_error
@@ -235,7 +236,7 @@ class GitRepository(Repository):
             cmd.extend(filenames)
         with self.lock:
             status = self.execute(cmd, merge_err=False)
-        return status != ""
+        return bool(status)
 
     def show(self, revision):
         """
@@ -482,7 +483,7 @@ class GitRepository(Repository):
             try:
                 self.execute(["fetch", "origin", *self.get_depth()])
             except RepositoryException as error:
-                if error.retcode == 1 and error.args[0] == "":
+                if error.retcode == 1 and not error.args[0]:
                     # Fetch with --depth fails on blank repo
                     self.execute(["fetch", "origin"])
                 else:
@@ -963,7 +964,7 @@ class GitMergeRequestBase(GitForcePushRepository):
                         params=params,
                         json=json,
                     )
-                except OSError as error:
+                except (OSError, HTTPError) as error:
                     report_error(cause="request")
                     raise RepositoryException(0, str(error))
 
@@ -1589,5 +1590,18 @@ class BitbucketServerRepository(GitMergeRequestBase):
             "post", credentials, pr_url, json=request_body
         )
 
+        """
+        Bitbucket Server will return error if PR already exists. The push
+        method in parent class will push changes to the correct fork or
+        branch, and always call this create_pull_request method after. If PR
+        exist already just do nothing because Bitbucket will automatically
+        update the PR if the from ref is updated.
+        """
         if "id" not in response:
+            pr_exist_message = (
+                "Only one pull request may be open "
+                "for a given source and target branch"
+            )
+            if pr_exist_message in error_message:
+                return
             raise RepositoryException(0, f"Pull request failed: {error_message}")