+++ weblate/vcs/mercurial.py
@@ -68,7 +68,7 @@ class HgRepository(Repository):
         config.read(filename)
         if config.has_option(section, option):
             result = config.get(section, option)
-        return result  # noqa: RET504
+        return result
 
     def set_config(self, path, value):
         """Set entry in local configuration."""
@@ -165,7 +165,7 @@ class HgRepository(Repository):
         if filenames:
             cmd.extend(filenames)
         status = self.execute(cmd, needs_lock=False)
-        return status != ""
+        return bool(status)
 
     def _get_revision_info(self, revision):
         """Return dictionary with detailed revision information."""